Rethinking testability: extract pure functions
